Test performance improvements.
Fix tests that were broken under xdist. uv run pytest -n 4 (or auto) now works
as expected.

* Initialize a single Ray cluster across tests.

This re-uses the cluster across multiple testing jobs. There's a tradeoff that
a previous module might leave Ray in a dirty state, but we should be able to
rely on Ray's own isolation to deal with this.  When running with xdist,
construct the cluster on worker 0 (or master if no parallelism is enabled).
This reduces the risk of "multiple Ray clusters found" race condition when
initializing the Ray client.

We need to disable 3 tests in xdist mode due to Ray weirdness.

* Switch runtime initialization for dedupe to use `uv` for initialization. i

This speeds up duplicate tests significantly and is ~50% faster to initialize.
It's still slow.

* Fix deprecated findAll -> find_all

Original-Commit: https://github.com/marin-community/marin/commit/9aaaa58d1865eb63cf706bac4a3fbaaccc968f0b